requirements.txt
# --- Core Python and API Frameworks ---
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-multipart==0.0.9
pydantic==2.11.3
pydantic-settings==2.11.0

# --- AI / LLM / Cloud Integrations ---
openai==1.51.0
assemblyai==0.34.0
mcp==1.16.0

# --- CV / Audio / NLP ---
mediapipe==0.10.15
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4
scipy==1.13.1
torch==2.4.1
torchaudio==2.4.1
transformers==4.46.0
sentence-transformers==3.2.1
pinecone==7.3.0

# --- Utility Libraries ---
pdfplumber==0.11.4
pandas==2.2.3
tqdm==4.66.5
requests==2.32.3
aiohttp==3.10.9
weasel==0.4.1
tenacity==9.0.0
httpx==0.27.2
beautifulsoup4==4.12.3

# --- Dev & Debug Tools ---
ipykernel==6.29.5
jupyterlab==4.2.5
python-dotenv==1.0.1
